+++ imio/dms/mail/browser/table.py
@@ -7,10 +7,13 @@ from collective.task import _ as _task
 from html import escape
 from imio.dms.mail import _
 from imio.dms.mail import _tr
+from imio.dms.mail.utils import get_approval_annot
 from imio.helpers.adapters import NoEscapeLinkColumn
 from imio.helpers.content import base_getattr
 from plone import api
 from Products.CMFPlone.utils import safe_unicode
+from Products.Five import BrowserView
+from Products.statusmessages.interfaces import IStatusMessage
 from z3c.table.column import Column
 from z3c.table.table import Table
 from zope.annotation.interfaces import IAnnotations
@@ -187,3 +190,118 @@ class PersonnelTable(Table):
     @CachedProperty
     def values(self):
         return self.results
+
+
+class FileNameColumn(Column):
+    """Column displaying file name."""
+
+    header = _(u"File name")
+
+    def renderCell(self, item):
+        return item.title
+
+
+class SignerColumn(Column):
+    """Column with checkboxes for each signer."""
+
+    def __init__(self, context, request, table, signer):
+        super(SignerColumn, self).__init__(context, request, table)
+        self.userid, self.signer = signer
+        if self.signer:
+            self.header = self.signer["name"]
+
+    def renderCell(self, item):
+        row_id = item.title
+        order = self.signer["order"]
+        checked = False
+        if item.UID() in self.table.annot["files"]:
+            checked = self.table.annot["files"].get(item.UID())[order]["status"] == "a"
+        name = "approvals.%s.%s" % (item.UID(), self.userid)
+        checked_attr = 'checked="checked"' if checked else ''
+        return u'<input type="checkbox" name="%s" %s />' % (name, checked_attr)
+
+
+class NoApprobationColumn(SignerColumn):
+    """Special column for 'No approbation'."""
+    header = _(u"* No validation")
+
+    def renderCell(self, item):
+        row_id = item.title
+        name = "approvals.%s.no_approval" % item.UID()
+        checked = item.UID() not in self.table.annot["files"]
+        checked_attr = 'checked="checked"' if checked else ''
+        return u'<input type="checkbox" name="%s" %s />' % (name, checked_attr)
+
+
+class ApprovalTable(Table):
+    """Table displaying approval state for admins."""
+
+    cssClassEven = u"even"
+    cssClassOdd = u"odd"
+    cssClasses = {'table': 'listing'}
+    sortOn = None
+
+    def __init__(self, context, request):
+        super(ApprovalTable, self).__init__(context, request)
+        self.annot = get_approval_annot(self.context)
+
+    def setUpColumns(self):
+        cols = []
+
+        # First column: file name
+        cols.append(FileNameColumn(self.context, self.request, self))
+
+        # Second column: no approbation
+        cols.append(NoApprobationColumn(self.context, self.request, self, (None, None)))
+
+        # Add approving columns
+        for signer in self.annot["users"].items():
+            col = SignerColumn(self.context, self.request, self, signer)
+            cols.append(col)
+
+        return cols
+
+
+class ApprovalTableView(BrowserView):
+    """Main view for approvals table."""
+
+    __table__ = ApprovalTable
+
+    def __init__(self, context, request):
+        super(ApprovalTableView, self).__init__(context, request)
+        self.table = self.__table__(context, request)
+
+    def __call__(self):
+        self.update()
+        self.handle_form()
+        return self.index()
+
+    def update(self):
+        self.table.results = self.context.listFolderContents({'portal_type': ['dmsommainfile', 'dmsappendixfile']})
+        self.table.update()
+
+    def handle_form(self):
+        form = self.request.form
+        save_button = form.get('form.button.Save', None) is not None
+        cancel_button = form.get('form.button.Cancel', None) is not None
+        if save_button and not cancel_button:
+            if not self.validate(form):
+                IStatusMessage(self.request).addStatusMessage(
+                    _(u"You cannot select approvings and no validation at the same time."), type='error')
+                return
+            # TODO Process form data here
+            IStatusMessage(self.request).addStatusMessage(
+                _(u"Changes saved."), type='info')
+
+    def validate(self, form):
+        """Validator: no_approval and signer's approval cannot both be checked."""
+        res = {}
+        for key in form.keys():
+            if not key.startswith("approvals."):
+                continue
+            prefix, file_uid, userid = key.split(".")
+            if file_uid not in res:
+                res[file_uid] = userid != "no_approval"
+            elif userid == "no_approval" and res[file_uid]:
+                return False
+        return True